requirements.txt
streamlit>=1.30.0
pillow>=9.0.0
numpy>=1.23.0
opencv-contrib-python-headless>=4.8.0
